requirements.txt
ale_py==0.7.5
gym==0.25.1
gym_minigrid==1.0.1
imageio==2.27.0
matplotlib==3.7.4
networkx==3.1
numpy==1.24.4
pygame
PyYAML==6.0.1
scikit_learn==1.3.0
torch==2.1.1
torch_geometric==2.4.0
